requirements.txt
python-telegram-bot
wikipedia